[Users] Update tests
